Basic Auth with hardcoded body
I couldn't find any mention in the HTTP specs about sending back a body with
what I assume is the 'protection space', and the acceptance test doesn't
force the GET /log, HEAD /requests etc to fail unless authorized.  I have presumed
that it is the case though.
Also, the AuthenticateAction feels like it should have some configuration values e.g
for the 'admin:hunter' details and the name of the 'realm'.